Employee,User service cleanup.

+++ src/main/java/com/javee/attendance/controllers/employee/EmployeeController.java
@@ -2,6 +2,7 @@ package com.javee.attendance.controllers.employee;
 
 import com.javee.attendance.controllers.BaseController;
 import com.javee.attendance.entities.Employee;
+import com.javee.attendance.entities.User;
 import com.javee.attendance.model.JWTUserDetails;
 import com.javee.attendance.repositories.EmployeeRepository;
 import io.swagger.annotations.Api;
@@ -9,7 +10,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
-import java.util.List;
 import java.util.Optional;
 
 @RestController
@@ -23,8 +23,15 @@ public class EmployeeController extends BaseController
 	@CrossOrigin
 	@RequestMapping( value = "/employee", method = RequestMethod.POST,
 			produces = "application/json", consumes = "application/json" )
-	public ResponseEntity<Employee> createEmpolyee( @AuthenticationPrincipal JWTUserDetails jwtUserDetails, @RequestBody Employee employee )
+	public ResponseEntity createEmpolyee( @AuthenticationPrincipal JWTUserDetails jwtUserDetails, @RequestBody Employee employee )
 	{
+		if (employee == null)
+			return generateBadRequestResponse();
+
+		Employee loggedInEmployee = getLoggedInEmployee( jwtUserDetails );
+		if (( !getLoggedInUserRole( jwtUserDetails ).equals( User.ROLE.ADMIN ) ) && ( loggedInEmployee == null || !employee.getId().equals( loggedInEmployee.getId() ) ))
+			return generateUnauthorizedResponse();
+
 		employee = employeeRepository.save( employee );
 		return generateOkResponse( employee );
 	}
@@ -32,30 +39,48 @@ public class EmployeeController extends BaseController
 	@CrossOrigin
 	@RequestMapping( value = "/employee/{id}", method = RequestMethod.GET,
 			produces = "application/json" )
-	public ResponseEntity<Employee> getEmployeeById( @AuthenticationPrincipal JWTUserDetails jwtUserDetails, @PathVariable( "id" ) Long id )
+	public ResponseEntity getEmployeeById( @AuthenticationPrincipal JWTUserDetails jwtUserDetails, @PathVariable( "id" ) Long id )
 	{
 		if (id == null || id == 0)
 			return generateBadRequestResponse();
 
+		Employee loggedInEmployee = getLoggedInEmployee( jwtUserDetails );
+		if (( !getLoggedInUserRole( jwtUserDetails ).equals( User.ROLE.ADMIN ) ) && ( loggedInEmployee == null || !id.equals( loggedInEmployee.getId() ) ))
+			return generateUnauthorizedResponse();
+
 		Optional<Employee> employeeOptional = employeeRepository.findById( id );
-		return generateOkResponse( employeeOptional.get() );
+
+		employeeOptional.ifPresent( this::generateOkResponse );
+
+		return generateNotFoundResponse();
 	}
 
 	@CrossOrigin
 	@RequestMapping( value = "/employees", method = RequestMethod.GET,
 			produces = "application/json" )
-	public ResponseEntity<List<Employee>> getEmployees( @AuthenticationPrincipal JWTUserDetails jwtUserDetails )
+	public ResponseEntity getEmployees( @AuthenticationPrincipal JWTUserDetails jwtUserDetails )
 	{
+		Employee loggedInEmployee = getLoggedInEmployee( jwtUserDetails );
+		if (( !getLoggedInUserRole( jwtUserDetails ).equals( User.ROLE.ADMIN ) ) && ( loggedInEmployee == null ))
+			return generateUnauthorizedResponse();
+
+		if (employeeRepository == null)
+			return generateNotFoundResponse();
+
 		return generateOkResponse( employeeRepository.findAll() );
 	}
 
 	@CrossOrigin
 	@RequestMapping( value = "/employee/{id}", method = RequestMethod.PUT,
 			produces = "application/json", consumes = "application/json" )
-	public ResponseEntity<Object> updateEmployee( @AuthenticationPrincipal JWTUserDetails jwtUserDetails, @RequestBody Employee employee, @PathVariable Long id )
+	public ResponseEntity updateEmployee( @AuthenticationPrincipal JWTUserDetails jwtUserDetails, @RequestBody Employee employee, @PathVariable Long id )
 	{
 		if (id == null || id == 0)
 			return generateBadRequestResponse();
+
+		Employee loggedInEmployee = getLoggedInEmployee( jwtUserDetails );
+		if (( !getLoggedInUserRole( jwtUserDetails ).equals( User.ROLE.ADMIN ) ) && ( loggedInEmployee == null || !id.equals( loggedInEmployee.getId() ) ))
+			return generateUnauthorizedResponse();
 
 		Optional<Employee> employeeOptional = employeeRepository.findById( id );
 		if (!employeeOptional.isPresent())
@@ -68,12 +93,17 @@ public class EmployeeController extends BaseController
 	@CrossOrigin
 	@RequestMapping( value = "/employee/{id}", method = RequestMethod.DELETE,
 			produces = "application/json" )
-	public ResponseEntity<Object> deleteEmployeeById( @AuthenticationPrincipal JWTUserDetails jwtUserDetails, @PathVariable( "id" ) Long id )
+	public ResponseEntity deleteEmployeeById( @AuthenticationPrincipal JWTUserDetails jwtUserDetails, @PathVariable( "id" ) Long id )
 	{
 		if (id == null || id == 0)
 			return generateBadRequestResponse();
 
+		Employee loggedInEmployee = getLoggedInEmployee( jwtUserDetails );
+		if (( !getLoggedInUserRole( jwtUserDetails ).equals( User.ROLE.ADMIN ) ) && ( loggedInEmployee == null || !id.equals( loggedInEmployee.getId() ) ))
+			return generateUnauthorizedResponse();
+
 		employeeRepository.deleteById( id );
+
 		return generateNoContentResponse();
 
 	}
